Simplified domain name from start_backup_for_client to start_backup

+++ urbackupserver/server_settings.cpp
@@ -451,7 +451,7 @@ std::string ServerSettings::generateRandomAuthKey(size_t len)
 	std::string rchars="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
 	std::string key;
 	std::vector<unsigned int> rnd_n=Server->getSecureRandomNumbers(len);
-	for(int j=0;j<len;++j)
+	for(size_t j=0;j<len;++j)
 		key+=rchars[rnd_n[j]%rchars.size()];
 	return key;
 }
